+++ acsAttractorAnalysisInTime.py
@@ -205,7 +205,7 @@ for timeFileID in range(0,tmpNOF):
 			vecX = np.array(lx)
 			vecY = np.array(ly)
 			# Compute coseno
-			tmpCos = np.dot(vecX,vecY) / (np.linalg.norm(vecX) * np.linalg.norm(vecY))
+			tmpCos = float(np.dot(vecX,vecY) / (np.linalg.norm(vecX) * np.linalg.norm(vecY)))
 			ANG_overallResMatrix[idx,idy] = np.arccos(tmpCos)  * 360.00 / 6.28
 			# HAMMING DISTANCE and EUCLIDEAN DISTANCE
 			tmpHD = 0
@@ -223,7 +223,7 @@ for timeFileID in range(0,tmpNOF):
 			vecX = np.array(lx)
 			vecY = np.array(ly)
 			# Compute coseno
-			tmpCos = np.dot(vecX,vecY) / (np.linalg.norm(vecX) * np.linalg.norm(vecY))
+			tmpCos = float(np.dot(vecX,vecY) / (np.linalg.norm(vecX) * np.linalg.norm(vecY)))
 			ANG_overallResMatrixNOINFLUX[idx,idy] = np.arccos(tmpCos)  * 360.00 / 6.28
 			# HAMMING DISTANCE and EUCLIDEAN DISTANCE
 			tmpHD = 0
